event: Merge request.go into event.go
